Update multiline parsing performance
My original logic for this program parsed text one line at a time.

I added support for embedded line-breaks much later, and I didn't focus on performance for it, to my great chagrin.

The logic for multiline CSV with embedded newlines needed to avoid backtracking and to avoid storing excess data in the work buffer.