Correct suggestions from Resharper
It's been a while since I did a code polish pass.  Now that we no longer support DotNet 2.0, we are free to use more modern patterns.  LINQ is now available in all versions of the DotNet framework that we support.  We can also just exclude the CSV datatable from those versions that don't support it (dot net standard 20).